Changed how mixins are implemented
