+++ src/client/client-wrapper.ts
@@ -1,12 +1,10 @@
 import * as grpc from 'grpc';
 import * as jsforce from 'jsforce';
+import { ContactAwareMixin } from './mixins/contact-aware';
 import { Field } from '../core/base-step';
 import { FieldDefinition } from '../proto/cog_pb';
-import { ContactMixin } from './contact-mixin';
 
-export interface ClientWrapper extends ContactMixin {}
-
-export class ClientWrapper {
+class ClientWrapper {
 
   // For now, only support Username and Password Login (OAuth2 Resource Owner Password Credential)
   public static expectedAuthFields: Field[] = [{
@@ -164,3 +162,17 @@ export class ClientWrapper {
     });
   }
 }
+
+interface ClientWrapper extends ContactAwareMixin {}
+applyMixins(ClientWrapper, [ContactAwareMixin]);
+
+function applyMixins(derivedCtor: any, baseCtors: any[]) {
+  baseCtors.forEach((baseCtor) => {
+    Object.getOwnPropertyNames(baseCtor.prototype).forEach((name) => {
+          // tslint:disable-next-line:max-line-length
+      Object.defineProperty(derivedCtor.prototype, name, Object.getOwnPropertyDescriptor(baseCtor.prototype, name));
+    });
+  });
+}
+
+export { ClientWrapper as ClientWrapper };
