+++ src/core/grpc-server.ts
@@ -2,7 +2,6 @@ import * as grpc from 'grpc';
 import { CogServiceService as CogService } from '../proto/cog_grpc_pb';
 import { Cog } from './cog';
 import { ClientWrapper } from '../client/client-wrapper';
-import { ContactMixin } from '../client/contact-mixin';
 
 const server = new grpc.Server();
 const port = process.env.PORT || 28866;
@@ -21,9 +20,7 @@ if (process.env.USE_SSL) {
   credentials = grpc.ServerCredentials.createInsecure();
 }
 
-const client = ContactMixin(ClientWrapper);
-
-server.addService(CogService, new Cog(client));
+server.addService(CogService, new Cog(ClientWrapper));
 server.bind(`${host}:${port}`, credentials);
 server.start();
 console.log(`Server started, listening: ${host}:${port}`);
